Fixes to allow opening recipes
This needs to be investigated further. When opening a subrecipe, the dialog needs more info (name, description), which isn't present in the recipe definition. Eric was using the sub_recipe_types to find it based on the revision and name. However, that's a foreign key pointer to latest database objects, so the revision could be different.