fix(recipes): fixed the sub-recipe behavior in graph components
* Fix template error, breaking page when clicked

* Move loading recipe data from jobs if statement

Not entirely sure if this is correct, but seems to work as expected. The chart is on the recipe dialog, but the chart options were being set if it was a job node.

* Fixes to allow opening recipes

This needs to be investigated further. When opening a subrecipe, the dialog needs more info (name, description), which isn't present in the recipe definition. Eric was using the sub_recipe_types to find it based on the revision and name. However, that's a foreign key pointer to latest database objects, so the revision could be different.

* Fix to check if the condition_interface is available

* Add link to sub recipe

* Conditionally apply link

* Watch for route param changes

* Fix broken tests